Support correlated subqueries (the doo doo way)
For each row of the input, replace all references to the outer query in
the subquery by their respective values and execute the subquery.

Note that subqueries that are not correlated where previously unnested
but are now executed once for each row of the input. A new optimization
will be needed to unnested the subquery when possible.