[WIP] Refactor subquery unnesting
A lot of branches are not implemented: among others, I need to be able to push a dependent join down a map.

The refactoring exposed flaws in the join reordering and other optimizations, as well as introducing brand new bugs.

Tests in regression have been commented, next steps is to fix them.